requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pyope==0.2.2
pydantic==2.1.1
shapely==2.0.6
python-dotenv
python-jose[cryptography]==3.3.0
supabase==1.0.5
h3==3.7.7
reverse_geocode==1.6.5
geojson-rewind==1.1.0
pyproj==3.6.1